# executor: update stats table row cache in batch