<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #62020

Problem Summary:

Previously, the stats row cache is updated one by one for each table. If each update takes 1ms, 1000 tables takes 1s to complete.

### What changed and how does it work?

Update the stats table row cache in batch.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
  sysbench oltp_common --mysql-port=4000 --mysql-user=root --mysql-host=127.0.0.1 --mysql-db=test --threads=300 --db-driver=mysql --tables=1000 --table-size=1 prepare
  
  Before this PR:
  ```
  mysql> trace select * from information_schema.tables where table_schema = 'test';
  +-------------------------------------------+-----------------+--------------+
  | operation                                 | startTS         | duration     |
  +-------------------------------------------+-----------------+--------------+
  | trace                                     | 11:12:06.567197 | 2.148556268s |
  |   ├─session.ExecuteStmt                   | 11:12:06.567202 | 346.749µs    |
  |   │ ├─executor.Compile                    | 11:12:06.567215 | 257.758µs    |
  |   │ │ ├─planner.Preprocess                | 11:12:06.567216 | 10.281µs     |
  |   │ │ └─planner.Optimize                  | 11:12:06.567244 | 221.328µs    |
  |   │ └─session.runStmt                     | 11:12:06.567476 | 53.028µs     |
  |   ├─*executor.MemTableReaderExec.Next     | 11:12:06.567586 | 2.147435809s |
  |   └─*executor.MemTableReaderExec.Next     | 11:12:08.715046 | 2.663µs      |
  +-------------------------------------------+-----------------+--------------+
  8 rows in set (2.15 sec)
  ```
  
  After this PR:
  ```
  mysql> trace select * from information_schema.tables where table_schema = 'test';
  +-------------------------------------------+-----------------+-------------+
  | operation                                 | startTS         | duration    |
  +-------------------------------------------+-----------------+-------------+
  | trace                                     | 11:10:47.520650 | 30.419448ms |
  |   ├─session.ExecuteStmt                   | 11:10:47.520655 | 312.578µs   |
  |   │ ├─executor.Compile                    | 11:10:47.520667 | 231.332µs   |
  |   │ │ ├─planner.Preprocess                | 11:10:47.520669 | 5.266µs     |
  |   │ │ └─planner.Optimize                  | 11:10:47.520689 | 203.404µs   |
  |   │ └─session.runStmt                     | 11:10:47.520903 | 48.49µs     |
  |   ├─*executor.MemTableReaderExec.Next     | 11:10:47.520979 | 30.02937ms  |
  |   └─*executor.MemTableReaderExec.Next     | 11:10:47.551034 | 3.146µs     |
  +-------------------------------------------+-----------------+-------------+
  8 rows in set (0.03 sec)
  ```
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```